Refactor location of functions
